Update OptiWindNet to version 0.0.4 (#101)
* chore: bump optiwindnet version to 0.0.3 and make pyproject.toml include subpackages

* fix: overhaul of optiwindnet_wrap to work with v0.0.3

* fix: update demos to use path relative to demo script and optiwindnet v0.0.3

* wip: initial adjustment of optiwindnet tests to v0.0.3

Many more changes are needed since the two level wrapping was changed to a single level. Also the tests are not taking into account links between turbines that use detours or contours and, therefore, have lengths that do not correspond to the Euclidean distance between the coordinates.

* fix: add border features to site in optiwindnetCollection.site_from_inputs()

* chore: clean up confusing code comments

* chore: black-format .py files

* wip: attempt to fix CI errors by using main branch of statsmodels

* Change optiwindnet to v0.0.4 to fix error on MacOS

* fix: update obsolete mip_rel_gap to mip_gap

* fix: new data structure for optiwindnet-generated data

The cable lengths and loads are ordered in robust way. If the networks are the same, the length/load values must match independent of platform. To guarantee that the solution
network is the same, the optimum must be unique and there must be enough time to reach it

* lint with black

* fix: reference to outdated appsi_highs

* chore: improved docstring and type hinting

* fix: add missing import

* fix: test failure due to length_cables being an output but not having partials calculated

The OpenMDAO check_partials() expects partials for all declared outputs. Since partials wrt the length_cables array are not calculated in compute_partials(), the length_cables output was moved to discrete_outputs().

* test: add test with border to improve code coverage

* lint with black